etl/requirements.txt
pymongo
redis==5.0.7
pyyaml